Clean up console crate.
