+++ Cargo.toml
@@ -31,7 +31,6 @@ clean-path = "0.2.1"
 compact_str = { version = "0.9.0", default-features = false, features = [
     "serde",
 ] }
-console = "0.15.11"
 convert_case = "0.8.0"
 dirs = "6.0.0"
 futures = "0.3.31"
