+++ crates/app/src/lib.rs
@@ -2,8 +2,6 @@ mod app;
 mod app_error;
 pub mod commands;
 mod components;
-mod experiments;
-mod helpers;
 pub mod queries;
 mod session;
 pub mod systems;
