--- crates/console-reporter/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "moon_console_reporter"
-version = "0.0.1"
-edition = "2024"
-license = "MIT"
-description = "Customizable reporters for console output."
-homepage = "https://moonrepo.dev/moon"
-repository = "https://github.com/moonrepo/moon"
-publish = false
-
-[dependencies]
-moon_action = { path = "../action" }
-moon_common = { path = "../common" }
-moon_config = { path = "../config" }
-moon_console = { path = "../console" }
-moon_target = { path = "../target" }
-moon_time = { path = "../time" }
-miette = { workspace = true }
-starbase_console = { workspace = true }
-starbase_styles = { workspace = true }
-
-[lints]
-workspace = true
