--- crates/console-reporter/src/lib.rs
@@ -1,3 +1,0 @@
-// mod default_reporter;
-
-// pub use default_reporter::*;
