+++ crates/console/Cargo.toml
@@ -14,9 +14,7 @@ moon_common = { path = "../common" }
 moon_config = { path = "../config" }
 moon_target = { path = "../target" }
 moon_time = { path = "../time" }
-inquire = "0.7.5"
 miette = { workspace = true }
-parking_lot = "0.12.3"
 starbase_console = { workspace = true, features = ["ui"] }
 starbase_styles = { workspace = true }
 tracing = { workspace = true }
