+++ crates/console/src/theme.rs
@@ -1,0 +1,20 @@
+use starbase_console::ui::{ConsoleTheme, Style, style_to_color};
+
+pub fn create_console_theme() -> ConsoleTheme {
+    let mut theme = ConsoleTheme::branded(style_to_color(Style::Id));
+    let mut frames = vec![];
+
+    for i in 1..=20 {
+        if i == 20 {
+            frames.push("━".repeat(20));
+        } else {
+            frames.push(format!("{}╾{}", "━".repeat(i - 1), " ".repeat(20 - i)));
+        }
+    }
+
+    theme.progress_loader_frames = frames;
+    theme.progress_bar_filled_char = '━';
+    theme.progress_bar_unfilled_char = '─';
+    theme.progress_bar_position_char = '╾';
+    theme
+}
